refactor : 기존 데이터클래스들 패키지 이동 domain -> domain.data
